requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
BareNecessities==0.2.8
bcrypt==4.0.1
black==26.1.0
certifi==2026.1.4
cfgv==3.5.0
click==8.3.1
coverage==7.13.4
distlib==0.4.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.129.0
fastapi-limiter==0.2.0
filelock==3.24.3
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.16
idna==3.11
iniconfig==2.3.0
isort==8.0.0
Mail==2.1.0
Mako==1.3.10
MarkupSafe==3.0.3
mypy_extensions==1.1.0
nodeenv==1.10.0
packaging==26.0
passlib==1.7.4
pathspec==1.0.4
platformdirs==4.9.2
pluggy==1.6.0
pre_commit==4.5.1
psycopg2-binary==2.9.11
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyrate-limiter==4.0.2
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
python-discovery==1.0.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
pytokens==0.4.1
PyYAML==6.0.3
redis==7.2.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
virtualenv==21.0.0
